fixed remaining errors concerning npm version
