+++ example_crosslink.html
@@ -227,7 +227,7 @@
 
 	// $(document).ready(function () {
 
-	window.xiSPEC = new xispec.xiSPEC_wrapper({targetDiv: "xispec_wrapper"});
+	window.xiSPEC = xispec.createApp({targetDiv: "xispec_wrapper"});
 
 	window.xiSPEC.setData({
 		sequence1: sequence1,
